backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.9
PyPDF2==3.0.1
pydantic==2.5.3
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3

